+++ fbp_service.c
@@ -1,7 +1,6 @@
 // #include <firebird/fb_c_api.h>
 #include "php.h"
 #include "php_firebird.h"
-#include "php_firebird_includes.h"
 
 #include "service.h"
 #include "fbp_service.h"
