+++ service.c
@@ -2,7 +2,6 @@
 
 #include "php.h"
 #include "php_firebird.h"
-#include "php_firebird_includes.h"
 #include "zend_exceptions.h"
 
 #include "service.h"
