+++ var_info.c
@@ -1,6 +1,5 @@
 #include <ibase.h>
 #include "php.h"
-#include "php_firebird_includes.h"
 
 // zend_class_entry *FireBird_Var_Info_ce;
